moved dp, ddp outside of trainer
